migrate conf/upload.py functions to seed command group and register with the app for 'flask seed dev' support
